Create new file functionality with save file tweaks.
Creating a new file opens /modeler.html with the following message:
```javascript
{
  fileName: '',
  type: 'new'
}
```
This doesn't try to get an existing file from the backend and creates a new template.
The fileName is used for saving functionality.

Editing an existing file opens /modeler.html with the message:
```javascript
{
  fileName: '',
  type: 'edit'
}

Saving a file now sends a message back to the parent element so it can refresh itself to display any new files.